# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Knowledge Bank API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents with search, filters, sort, pagination
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: courseId, schema: { type: string, format: uuid } }
        - { in: query, name: courseCode, schema: { type: string } }
        - { in: query, name: lecturerName, schema: { type: string } }
        - { in: query, name: semester, schema: { type: string, enum: [fall, spring, summer] } }
        - { in: query, name: academicYear, schema: { type: integer } }
        - { in: query, name: materialType, schema: { type: string } }
        - { in: query, name: categoryId, schema: { type: string, format: uuid } }
        - in: query
          name: tagIds
          style: form
          explode: true
          schema: { type: array, items: { type: string, format: uuid } }
        - { in: query, name: dateFrom, schema: { type: string, format: date } }
        - { in: query, name: dateTo, schema: { type: string, format: date } }
        - { in: query, name: sort, schema: { type: string, enum: [newest, oldest, title, popularity], default: newest } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Page of documents
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentPage" }
...
  /documents/upload:
    post:
      operationId: uploadDocuments
      tags: [documents]
      summary: Upload one or more documents (multipart). Each file becomes its own Document.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
                courseId: { type: string, format: uuid }
                categoryId: { type: string, format: uuid }
                materialType: { type: string }
                semester: { type: string, enum: [fall, spring, summer] }
                academicYear: { type: integer }
                visibility: { type: string, enum: [public, restricted, private] }
                tagIds:
                  type: array
                  items: { type: string, format: uuid }
                # Optional title override; otherwise derived from filename
                title: { type: string }
                description: { type: string }
      responses:
        "201":
          description: Per-file results (partial success possible)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResult" }
  /documents/recent:
    get:
      operationId: listRecentDocuments
      tags: [documents]
      summary: The current user's recently viewed documents
      parameters:
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 20, default: 6 } }
      responses:
        "200":
          description: Recent documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Document" }
  /documents/suggestions:
    get:
      operationId: documentSuggestions
      tags: [documents]
      summary: Search suggestions for autocomplete
      parameters:
        - { in: query, name: q, required: true, schema: { type: string, minLength: 1 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 20, default: 8 } }
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DocumentSuggestion" }
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentDetail" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    patch:
      operationId: updateDocument
      tags: [documents]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDocumentRequest" }
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentDetail" }
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
  /documents/{id}/thumbnail:
    get:
      operationId: getDocumentThumbnail
      tags: [documents]
      summary: Stream a server-generated thumbnail (signed-URL pathway)
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: token, required: true, schema: { type: string } }
      responses:
        "200":
          description: Thumbnail bytes
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: No thumbnail for this document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /documents/{id}/preview-token:
    get:
      operationId: getDocumentPreviewToken
      tags: [documents]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Signed token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignedTokenResponse" }
  /documents/{id}/download-token:
    get:
      operationId: getDocumentDownloadToken
      tags: [documents]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Signed token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignedTokenResponse" }
  /documents/{id}/preview:
    get:
      operationId: previewDocument
      tags: [documents]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: token, required: true, schema: { type: string } }
      responses:
        "200":
          description: File stream
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /documents/{id}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: token, required: true, schema: { type: string } }
      responses:
        "200":
          description: File stream
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    UserSummary:
      type: object
      required: [id, email, displayName, roles, isActive, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        displayName: { type: string }
        roles:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        status:
          type: string
          enum: [ACTIVE, PENDING_APPROVAL, DISABLED]
        createdAt: { type: string, format: date-time }
...
    Course:
      type: object
      required: [id, code, title, lecturerName]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        title: { type: string }
        lecturerName: { type: string }
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
    Tag:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
    Document:
      type: object
      required:
        - id
        - title
        - description
        - materialType
        - visibility
        - status
        - uploader
        - createdAt
        - updatedAt
        - viewCount
        - commentCount
        - tags
        - permissions
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        course: { $ref: "#/components/schemas/Course" }
        category: { $ref: "#/components/schemas/Category" }
        materialType: { type: string }
        semester: { type: string, enum: [fall, spring, summer] }
        academicYear: { type: integer }
        visibility: { type: string, enum: [public, restricted, private] }
        status: { type: string, enum: [draft, published, archived] }
        uploader: { $ref: "#/components/schemas/UserSummary" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        viewCount: { type: integer }
        commentCount: { type: integer }
        tags:
          type: array
          items: { $ref: "#/components/schemas/Tag" }
        file: { $ref: "#/components/schemas/DocumentFileMeta" }
        thumbnailUrl:
          type: string
          description: >-
            Signed URL to a server-generated thumbnail when one exists.
            Issued by `assembleDocuments` after visibility checks; goes
            through the same signed-URL pathway as preview/download.
        fallbackIconType:
          type: string
          enum: [pdf, image, doc, slides, sheet, text, archive, unknown]
          description: >-
            Generic icon bucket the client renders when no thumbnail is
            available. Derived from the latest file's MIME type.
        permissions:
          $ref: "#/components/schemas/DocumentPermissions"
    DocumentPermissions:
      type: object
      description: >-
        Server-computed permission flags for the requesting user
        against this document. The frontend MUST use these flags
        (rather than role/uploader heuristics) to gate UI affordances
        — they encode the same course-aware logic the API enforces on
        write paths.
      required: [canView, canEdit, canDelete, canDownload, canComment]
      properties:
        canView: { type: boolean }
        canEdit: { type: boolean }
        canDelete: { type: boolean }
        canDownload: { type: boolean }
        canComment: { type: boolean }
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/Document"
    DocumentFileMeta:
      type: object
      required: [id, originalFilename, displayFilename, mimeType, sizeBytes, uploadedAt]
      properties:
        id: { type: string, format: uuid }
        originalFilename:
          type: string
          description: The exact filename as the user uploaded it.
        displayFilename:
          type: string
          description: The filename shown in lists; if the uploader already had a file with the same name it will be suffixed e.g. "notes (2).pdf". Use this for the rename notice.
        mimeType: { type: string }
        sizeBytes: { type: integer }
        uploadedAt: { type: string, format: date-time }
        checksum: { type: string }
        extractedMetadata:
          $ref: "#/components/schemas/ExtractedFileMetadata"
    ExtractedFileMetadata:
      type: object
      required: [hasExtractedText]
      description: >-
        Server-side metadata pulled from the uploaded file on ingest
        (task #27). Every field is optional — extraction may fail
        per-file without failing the upload.
      properties:
        pageCount: { type: integer, minimum: 0 }
        detectedTitle: { type: string }
        author: { type: string }
        imageWidth: { type: integer, minimum: 0 }
        imageHeight: { type: integer, minimum: 0 }
        hasExtractedText:
          type: boolean
          description: True when extracted text exists for full-text search (task #28).
    DocumentSuggestion:
      type: object
      required: [id, title]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        courseCode: { type: string }
        materialType: { type: string }
    DocumentPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Document" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    UpdateDocumentRequest:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        courseId: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        materialType: { type: string }
        semester: { type: string, enum: [fall, spring, summer] }
        academicYear: { type: integer }
        visibility: { type: string, enum: [public, restricted, private] }
        status: { type: string, enum: [draft, published, archived] }
        tagIds:
          type: array
          items: { type: string, format: uuid }
    UploadResult:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items: { $ref: "#/components/schemas/UploadFileResult" }
    UploadFileResult:
      type: object
      required: [originalFilename, success]
      properties:
        originalFilename: { type: string }
        success: { type: boolean }
        document: { $ref: "#/components/schemas/Document" }
        error: { type: string }
        errorCode:
          type: string
          description: >-
            Machine-readable failure reason. Known values:
            `disallowed_mime`, `mime_mismatch`, `storage_quota_exceeded`,
            `duplicate_file`, `upload_failed`.
        duplicateOfDocumentId:
          type: string
          format: uuid
          description: >-
            Set when `errorCode === "duplicate_file"` — id of the existing
            document this upload matched by SHA-256.
        duplicateOfTitle:
          type: string
          description: >-
            Set when `errorCode === "duplicate_file"` — title of the
            existing document this upload matched by SHA-256.
...
    SignedTokenResponse:
      type: object
      required: [token, expiresAt, url]
      properties:
        token: { type: string }
        expiresAt: { type: string, format: date-time }
        url: { type: string, description: "Fully-qualified URL with token already attached" }
...