# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Knowledge Bank API
 * OpenAPI spec version: 0.2.0
 */
import type { Category } from "./category";
import type { Course } from "./course";
import type { DocumentFallbackIconType } from "./documentFallbackIconType";
import type { DocumentFileMeta } from "./documentFileMeta";
import type { DocumentPermissions } from "./documentPermissions";
import type { DocumentSemester } from "./documentSemester";
import type { DocumentStatus } from "./documentStatus";
import type { DocumentVisibility } from "./documentVisibility";
import type { Tag } from "./tag";
import type { UserSummary } from "./userSummary";

export interface Document {
  id: string;
  title: string;
  description: string;
  course?: Course;
  category?: Category;
  materialType: string;
  semester?: DocumentSemester;
  academicYear?: number;
  visibility: DocumentVisibility;
  status: DocumentStatus;
  uploader: UserSummary;
  createdAt: Date;
  updatedAt: Date;
  viewCount: number;
  commentCount: number;
  tags: Tag[];
  file?: DocumentFileMeta;
  /** Signed URL to a server-generated thumbnail when one exists. Issued by `assembleDocuments` after visibility checks; goes through the same signed-URL pathway as preview/download. */
  thumbnailUrl?: string;
  /** Generic icon bucket the client renders when no thumbnail is available. Derived from the latest file's MIME type. */
  fallbackIconType?: DocumentFallbackIconType;
  permissions: DocumentPermissions;
}