# lib/api-zod/src/generated/types/uploadFileResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Knowledge Bank API
 * OpenAPI spec version: 0.2.0
 */
import type { Document } from "./document";

export interface UploadFileResult {
  originalFilename: string;
  success: boolean;
  document?: Document;
  error?: string;
  /** Machine-readable failure reason. Known values: `disallowed_mime`, `mime_mismatch`, `storage_quota_exceeded`, `duplicate_file`, `upload_failed`. */
  errorCode?: string;
  /** Set when `errorCode === "duplicate_file"` — id of the existing document this upload matched by SHA-256. */
  duplicateOfDocumentId?: string;
  /** Set when `errorCode === "duplicate_file"` — title of the existing document this upload matched by SHA-256. */
  duplicateOfTitle?: string;
}