# lib/api-zod/src/generated/types/userSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Knowledge Bank API
 * OpenAPI spec version: 0.2.0
 */
import type { UserSummaryStatus } from "./userSummaryStatus";

export interface UserSummary {
  id: string;
  email: string;
  displayName: string;
  roles: string[];
  isActive: boolean;
  status: UserSummaryStatus;
  createdAt: Date;
}